refactor: move ShortcutKey repository
