+++ src/BookmarkEditor/component.tsx
@@ -1,9 +1,10 @@
 import './component.css'
-import { removeBookmark, useShortcutMap } from '../Bookmarks/repository'
 import { Bookmark } from '../Bookmarks/model'
 import { EditingBookmark } from './model'
 import { FC } from 'react'
+import { removeBookmark } from '../Bookmarks/repository'
 import { shortcutKeyOf } from '../ShortcutKey/model'
+import { useShortcutMap } from '../ShortcutKey/repository'
 
 type BookmarkEditorComponentProps = {
   editingBookmark?: EditingBookmark
