+++ src/Bookmarks/repository.ts
@@ -1,6 +1,5 @@
 import { Bookmark, BookmarkFolder, BookmarkFolderID, chromePages } from './model'
 import { useEffect, useState } from 'react'
-import { ShortcutMap } from '../ShortcutKey/model'
 import { useChromeStorage } from '../infrastructure/chromeStorage'
 
 export const useBookmarkFolders = () => {
@@ -85,17 +84,3 @@ export const useCollapsedBookmarkFolderIDs = () =>
       }
     },
   })
-
-export const useShortcutMap = (): [ShortcutMap, (newMap: ShortcutMap) => void] => {
-  const [entries, setEntries] = useChromeStorage<[string, string][]>({
-    areaName: 'sync',
-    key: 'v3.shortcutKeyMap',
-    initialValue: [],
-    assertType: (value: unknown) => {
-      if (!Array.isArray(value)) {
-        throw new Error('value is not array')
-      }
-    },
-  })
-  return [new ShortcutMap(entries), (newMap: ShortcutMap) => setEntries(newMap.serialize())]
-}
